Separate codex to glossary and annotations suites
