+++ .github/workflows/playwright.yml
@@ -10,23 +10,30 @@ jobs:
       test_name: articles navigation
       device_type: "Desktop Chrome"
 
-  e2e_desktop_codex:
-    name: Desktop Chrome - Codex
+  e2e_desktop_glossary:
+    name: Desktop Chrome - Glossary
     uses: ./.github/workflows/e2e.yml
     with:
-      test_name: Codex
+      test_name: Glossary
       device_type: "Desktop Chrome"
 
-  e2e_mobile:
-    name: Mobile Chrome
+  e2e_desktop_annotations:
+    name: Desktop Chrome - Annotations
     uses: ./.github/workflows/e2e.yml
     with:
-      test_name: navigation.mobile
-      device_type: "Mobile Chrome"
+      test_name: Annotations
+      device_type: "Desktop Chrome"
 
-  e2e_mobile_codex:
-    name: Mobile Chrome - Codex
-    uses: ./.github/workflows/e2e.yml
-    with:
-      test_name: codex
-      device_type: "Mobile Chrome"
+  # e2e_mobile:
+  #   name: Mobile Chrome
+  #   uses: ./.github/workflows/e2e.yml
+  #   with:
+  #     test_name: navigation.mobile
+  #     device_type: "Mobile Chrome"
+
+  # e2e_mobile_codex:
+  #   name: Mobile Chrome - Codex
+  #   uses: ./.github/workflows/e2e.yml
+  #   with:
+  #     test_name: codex
+  #     device_type: "Mobile Chrome"
